Move replicator_startup_test to integration test set
Sometimes fails when run multithreaded with other tests.
